Added comment handling to MsgParse and initial MsgGen support.
Refactored the built-in ROS message types to the MsgTypes module.

Implemented some big pieces of MsgGen (generating a Haskell source
file from a Msg value), including the data type definition and a
Binary instance (though this is untested).

The holdup at the moment is that I don't know how to do the
BinaryIter instance for the custom type. One problem is that
BinaryIter isn't compositional at all at the moment: BinaryIter
instances of constituent values are no use to aggregate types becaues
the continuation (More) variant is monomorphic.

This means that either the BinaryIter instance must be built-up
by hand for every new type, or that the BinaryIter TypeClass should
be changed to make Emit return the remaining ByteString.