Fixed hierarchical module names and sanitized field names.
Message modules live in Ros.Package and must be defined as such
(e.g. Ros.Std_msgs.String).

Some field names are not valid in Haskell, e.g. "data", so are
prefixed with an underscore when generating Haskell code.

Haskell types in msg modules are now qualified to avoid conflicts. For
example, the Prelude's String type clashes with the ROS String message
type, so Prelude is a qualified import.