Added qualified user type module imports.
Fixed bug in HasHeader instance generation. Only the sequence number
can be set, not the entire header field.

This includes non-built-in standard types such as Header, which
requires additional extra support for the HeaderSupport module (where
the HasHeader class is defined).

The general idea is to qualify a lot of names in message definitions
to avoid conflicts. This shouldn't affect the usage of these modules
as the end-user will have to manage his or her own imports
appropriately.