Fixed integration with Data.Stream. Improved Vector serialization performance. Improved network handshake robustness. Improved cabal integration.
Data.Stream uses lazy pattern matching which would cause a
publisher with no subscribers to spin as fast as possible and
leak memory. Adding a strictness annotation to the publisher's
publication loop in RosTcp fixed this.

Improved pointer swapping between ByteStrings and
Storable Vectors.

Explicitly make sure that we receive as many bytes as we need on
socket recv calls.

Further customization of Setup.hs to force the desired --bindir
and --libdir options to cabal install, as well as automatic
appending of additional hsSourceDirs containing the current
package's message dependencies.