Defined interpolation on Streams and wrote a quick test.
The FastAndSlow module now has a definition for interpolation that
is blind to any concrete notion of time stamps, or even literal ordering.

Instead, this interpolation uses the disjunction of two streams (which
produces a value whenever either component stream produces a value) to
define an implicit ordering. Consecutive elements of the first stream
are used to bracket elements of the second stream, all of which is
handed off to a user-supplied interpolation function that returns a
list of new values that are ultimately streamed out.