Working on improving advertiseIO.
Things are in a bit of a sloppy state at the moment as I try
to work out the right strategy.

The configuration that triggered the issue was mapping a function
that returned an IO action over a Stream, then advertiseIO'ing
that action. This resulted in a deadlock. The current situation is
heavy on threads and not well factored. That said, the immediate
problem seems to have been addressed.