Added interruptible to StreamCombinators.
When producing a stream by flattening the elements of another
(e.g. converting a Stream [a] to Stream a), one may wish to abandon
the processing of an earlier element whenever a new value arrives on
the original stream. The interruptible function allows for just this
usage. The original Stream's Foldable values are traversed and fed
into the new Stream until a new value arrives on the original Stream,
at which point the traversal of that new value begins.

- Added an example program (Interruptible.hs) that demonstrates usage.