Finished first version of the flatness analyzer.
Flat message types are given a Storable instance, thus enabling
collections of these types to be stored in Vectors in other message
types. The motivating case for this is the PointCloud message that
contains an array of Point32 values. A Point32 is just three Float32s,
and so can be made an instance of Storable.

Also added support for deprecated message field types "byte" and
"char" which have been superceded by uint8 and int8, respectively.