Fixed a bug in generated peek method of Storable instances.
Handling of flattened array types does not work. The get and put
RosBinary methods need to know how to get and put each field. I should
do a single pass over all a Msg's fields, annotating each one with an
indicator of whether or not it is flat. Then generate the data
declaration using that, and generate the relevent instances from it,
too.

A second issue is that FixedArrays are not considered as flat, even
though they could be if the nested type is flat. This may not be
possible as it would require some finesse to ensure that values of the
type always have vector fields of the specified length. Perhaps this
should be set aside.