Improved caching performance of package message generation.
"roshask dep" now runs somewhat faster, though still not great. I
believe a significant part of the slowdown is due to md5 generation,
which I could speed up by pulling it into roshask at the risk of
losing compatibility with the canonical rospy implementation

There is a new module, Ros.Logging, that can not be built until
roshask is installed and run over a base ROS installation. This can be
a first candidate for a roshask-util package. The issue is that this
module can not be compiled without ROS.