Improved Logging with TH for file name and line number annotations.
Import Ros.Logging in Example5 to see if it's working.

An issue is that I don't currently have the node name set in the log
messages. This can not be done statically, and so would require that
the caller of a logging function live in some monad. Perhaps the Node
monad? But then how can I fmap such a function over a Stream?