Added ROS message MD5 sum generation to roshask.
The advantage of this is that message generation is now much faster
than before as we don't need to do one dependency analysis to generate
the Haskell source and then a second for the roslib Python tools to do
MD5 generation. Now we can generate all the messages in a package very
quickly. The actual MD5 calculation is farmed out to OpenSSL, but we
make use of a single parsing pass over the raw .msg files.

Added an orphan NFData instance for Data.Vector.Storable.Vector. It is
unfortunate that this is an Orphan, but it will be needed to deepseq
ROS messages.