Added imperative style logging to /rosout.
In order to make logging available in the roshask package, I included
copies of the generated Haskell code for the Header and Log message
types. This boot strapping process may be controled by the "logging"
flag given to cabal. cabal install -f-logging will build roshask
without logging so that the the roshask executable can be used to
generate Log.hs and Header.hs. These files can then be moved into the
source tree in Ros/Core (and have their module names suitably
modified).

Logging is a grossly imperative procedure at the moment. It lives in
IO and uses mutable global state in order to funnel all log messages
into the /rosout Topic. While the design can certainly be improved,
this style of printf debugging is imperative in spirit, so perhaps
doing everything in IO with side effects doesn't signal the end of the
world.