Fixed interruptible with lazier tee.
Added a new implementation of tee that doesn't force a value from the
source Topic until one of the two teed Topics demands it. This is
critical to make composition with the interruptible combinator
work. The old version is still present as teeEager since it should
perform better if all needed laziness is granted by inherent IO
delays.

Removed toList/fromList variation from the Turtle example.