Fixed bug with Bool arrays.
While the RosBinary instance handles serialization of Haskell Bool
values to and from a single byte, the Storable instance for Bool uses
four bytes. A problem then occurs when we want to use a Storable Vector
to hold boolean values. If we receive four bytes for an array of boolean
values from a peer, that is four distinct booleans, but the Storable
instance for Bool wants the Vector of length four to use 16 bytes.

Instead, we simply map the Ros bool type to a Haskell Word8 for now.

Fixes #15