Bumped haxr dependency to address roscpp compat.
Some communication with roscpp nodes was causing trouble for roshask due
to an implicit string type annotation in XML-RPC. According to the spec,
an XML-RPC value such as "<value></value>" should be interpreted as an
empty string. haxr was instead looking for
"<value><string></string></value>" for such values.

I patched haxr to accept unwrapped text as a string value.

Fixes #14